+++ app/screens/channel/channel_nav_bar/channel_title/channel_display_name/index.tsx
